Correction of inspect command and print edges command

+++ madventure/madventure/application/system/levelsystem.h
@@ -16,7 +16,7 @@ public:
   shared_ptr<Level> NewLevel(int columns, int row);
   vector<int> GetNeighboringNodeIndices(shared_ptr<Area> area);
   void setWorld(shared_ptr<World> world) { world_ = world; }
-  static double RelativeVectorAngle(const Vector2D alpha, const Vector2D beta);
+  static int RelativeVectorAngle(const Vector2D alpha, const Vector2D beta);
 private:
   shared_ptr<World> world_;
 };
